% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coeffpml}
\alias{coeffpml}
\title{coeffpml}
\usage{
coeffpml(u1, v1, u2, v2, amin, parp, parh, n, na, mf)
}
\arguments{
\item{u1}{Ranks of x-values in subregion}

\item{v1}{Ranks of x-values in subregion}

\item{u2}{ranks of y-values in subregion}

\item{v2}{ranks of y-values in subregion}

\item{amin}{minimum fraction of sample items in a subregion}

\item{n}{total number of sample items}

\item{na}{number of data in subregion}

\item{mf}{int}
}
\value{
NumericVector
}
\description{
Computing whatever
}
\details{
Some details
}
\author{
Eckhard Liebscher
}
\keyword{internal}
