% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packdepcoeff.R
\name{kendrm}
\alias{kendrm}
\title{Multivariate Kendall regression coefficient}
\usage{
kendrm(x,y,direction=NULL,out=0,outsd=TRUE,eps=0.95)
}
\arguments{
\item{x}{data matrix of regressor variables}

\item{y}{data matrix of the response vector}

\item{direction}{vector of length d (d is number of regressors),
value 1 refers to regressors leading to increasing y whenever this regressor increases,
value -1 refers to regressors leading to decreasing y whenever this regressor increases.
If direction=NULL, then all coefficients are computed.}

\item{out}{value 1: full output, value 0: reduced output of one coefficient
that is largest in absolute value, value 2: only the largest coefficient}

\item{outsd}{logical. If TRUE, the estimated standard deviation and confidence intervals
are evaluated.  If FALSE, the evaluation is suppressed}

\item{eps}{real value, confidence level.}
}
\value{
A list of the coefficients for several directions with components:
  \describe{
    \item{dcoeff}{Kendall regression coefficient}
    \item{dir}{direction vector}
    \item{Pxx}{fraction of \eqn{X \le \breve{X}} where \eqn{\breve{X}} has the same distribution as \eqn{X}}
    \item{sd}{standard deviation}
    \item{conf}{confidence interval}
  }
}
\description{
The function kendrm evaluates the multivariate Kendall regression coefficient.
It describes how well the response vector y can be fitted by a function of the regressor variables
which increases or decreases as the regressors increases.
}
\examples{
library(faraway)
y<- fat[,c("brozek","siri")]
x<- fat[,c("weight","adipos","abdom","hip")]
kendrm(x,y,direction=NULL,out=1,outsd=FALSE)
kendrm(x,y,direction=NULL,out=0,outsd=TRUE)
}
\references{
Eckhard Liebscher (2026). Kendall regression coefficient revisited.
}
