% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packdepcoeff.R
\name{kendtaum}
\alias{kendtaum}
\title{Multivariate Kendall's tau}
\usage{
kendtaum(x,outsd=TRUE,eps=0.95)
}
\arguments{
\item{x}{data matrix of regressor variables}

\item{outsd}{logical. If TRUE, the estimated standard deviation and confidence intervals
are evaluated.  If FALSE, the evaluation is suppressed}

\item{eps}{real value, confidence level.}
}
\value{
A list of the coefficients for several directions with components:
  \describe{
    \item{dcoeff}{multivariate Kendall's tau}
    \item{sd}{standard error}
    \item{conf}{confidence interval}
  }
}
\description{
The function kendtaum evaluates the multivariate Kendall's tau coefficient.
It describes the dependence of the variables in the data matrix
}
\examples{
library(MASS)
data <- gilgais
kendtaum(data[,1:4],outsd=TRUE)
}
\references{
Schmid, F.; Schmidt, R.; Blumentritt, T.; Gaißer, S.; Ruppert, M. (2010).
Copula-based measures of multivariate association. in F. Durante, W. Härdle,
P. Jaworski, T. Rychlik (eds.) Copula Theory and Its Applications. Springer Berlin, 2010.
}
