% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packdepcoeff.R
\name{spearrs}
\alias{spearrs}
\title{Spearman regression coefficient for split domains}
\usage{
spearrs(x,y,splitp=NULL)
}
\arguments{
\item{x}{data matrix of regressor variables with two columns,}

\item{y}{data vector of the target variable}

\item{splitp}{vector of length 2 of the splitting points,
If p1 is the first component of this vector, then the point splits the domain of the
first regressor into a left region of fraction p1 of data items and a right region
of the remaining data items. The same is done for the second regressor. As the
result we obtain 4 subregions of the regressor domain. default=c(0.5,0.5)}
}
\value{
A list of Spearman regression coefficients for the 4 split regions
(11=left lower region, 12=left upper region...) with components:
  \describe{
    \item{dcoeff++}{split coefficient ++}
    \item{dcoeff+-}{split coefficient +-}
    \item{totalcoeff}{total coefficient}
    \item{directions}{optimal directions for the several split regions}
  }
direction ++ means that y increases whenever both regressors increases
direction +- means that y increases whenever the first regressor increases and the
other regressor decreases..etc.
}
\description{
The function spearrs evaluates the multivariate Spearman regression coefficient
for two regressors and split regressor region.
It describes how well the response variable can be fitted in each split region
by a function which increases or decreases as the regressors increase.
}
\examples{
library(MASS)
data<- gilgais
spearrs(data[,1:2],data[,3],splitp=c(0.4,0.6))
}
\references{
Eckhard Liebscher (2021). On a multivariate version of Spearman's
correlation coefficient for regression: Properties and Applications. Asian Journal
of Statistical Sciences 1, No. 2, 123-150.
}
